Move out ticket and server connection logic out of main.cpp
