+++ client/src/exceptions-base.cpp
@@ -1,0 +1,9 @@
+#include "exceptions-base.hpp"
+
+ExceptionWithErrorMsg::ExceptionWithErrorMsg(const std::string &error)
+    : error_msg(error)
+    {}
+
+std::string ExceptionWithErrorMsg::get_error_msg() const {
+    return error_msg;
+}
