+++ client/src/ticket.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <string>
+#include "exceptions-base.hpp"
+
+class TicketError : public ExceptionWithErrorMsg {
+    using ExceptionWithErrorMsg::ExceptionWithErrorMsg; //inherit constructor
+};
+
+class Ticket {
+private:
+    std::string ticket_token;
+    std::string totp_secret;
+public:
+    Ticket(const std::string &ticket_token, const std::string &totp_secret);
+
+    std::string get_ticket_string() const;
+};
